(Restructured code. Added car sprite assets. Added tool classes.)
Project structure has been made a little more conventional. The main class no longer nests anonymous functions. Instead it defines an initialisation function (`init()`) which loads all assets, and a `postLoadEverything()` function that instantiates the game class.

All game classes are stored in the `game` package. The main game class is now `SpeedRace` and it creates and positions a new instance of `PlayerCar` on the screen.

However, one thing is new here - the backbuffer. Essentially, this is an image that can be drawn to the same way the framebuffer can. Typically the backbuffer is drawn to by all other elements, and once all of the draw calls are executed the filled backbuffer is passed on to the framebuffer to display the final result. Here the backbuffer is created with the 240 x 320 rendering resolution in mind and it is scaled up or down to the framebuffer's dimensions (which depend on the window, HTML5 canvas, or other render target's size). It's nice to establish a baseline for rendering like this because it helps manage expectations of what the final result will look like.

The usual "tool" classes, like `Object`, `Entity`, `Sprite`, and so on, are now packaged under `game.tools`. They are slightly different than in earlier iterations (see kha-pong, kha-astro-race, kha-basketball), but only marginally so. For instance, `Object` now defines `isActive` and `isVisible` flags as well as `kill()` and `revive()` methods to directly unset or set them. These can be helpful when implementing mechanics such as object recycling.