(Implemented road as group of road segments. Implemented vertical movement.)
Speed Race is one of the first games to use scrolling sprites. The player car could only move horizontally on the road, and the world around it would scroll vertically giving the impression of movement. Having other cars scroll on the screen is trivial and a similar visual effect was implemented in the Astro Race clone (kha-astro-race). Scrolling the road, however, is a little trickier.

The road is treated as a group of road segment sprites. The number of segments is given as the ratio of the backbuffer height to the segment graphic height. Two extra segments are added to the group for a more seamless scrolling experience. When a segment moves past the "bottom" of the backbuffer it is repositioned at the top. An offset needs to be taken into account due to the speed at which the segments are scrolling for proper repositioning.