(Implemented horizontal movement for player car.)
The player car can only move horizontally. The class is therefore given an `updateMovement()` method that takes the velocity and direction for arguments. This method is called on the 'a' or 'd' key press by the key handlers in the `SpeedRace` game class. Nothing out of the ordinary, really.

Only thing that might stand out is the use of static variables to limit horizontal movement to the area of the game occupied by the road. Because all car movement will be horizontally bound in this fashion, there is no need to check for some sort of collision against the road or the sides of the road. Instead, minimum and maximum x-coordinates are set which limit movement of any Car subclass accordingly.