(Added simple score display.)
Score increases the faster the player drives, therefore it is calculated from the current velocity and maximum velocity. An 0.1 score modifier was chosen to control the rate at which the score increases.

Because the game is rendered to a backbuffer and scaled to the framebuffer's dimensions, text can end up muddy if kept on the backbuffer. This is why the score text is drawn to the framebuffer after the backbuffer has been scaled. It should be noted that the scale transformation still exists at that point and needs to be replaced with an identity matrix, otherwise the text will also be scaled in the same way the backbuffer was.