(Added NPC cars. Added `Pool` as an extension of `Group` that limits number of elements and allows for recycling. First version of clone is finished.)
Computer controlled cars (`NpcCar`) are sporadically created to thwart the player's progress. Because they scroll vertically across the screen they can be recycled and reused once they're no longer visible. To help with this, the `Pool` class was created. It is an extension of `Group` which limits the amount of available elements in it and allows for recycling of the inactive and invisible ones. Every time an `NpcCar` moves across the screen its `kill()` method is called to make it invisible and inactive. Once another `NpcCar` is required, it is recycled from the pool by picking up the first 'killed' car. This car is revived with `revive()`, reinitialised, and reintroduced to the game world.

The `isRunning` flag was introduced to `SpeedRace` to control the flow of the game. It indicates whether the game is running (post `startGame()` call) or not (post `stopGame()` call). Nothing special about it as it does nothing more than the job it is assigned to do - branch the code depending on whether the game is being played or not.